Refactor delete invoice functionality

+++ nextjs-dashboard/app/lib/actions.ts
@@ -3,7 +3,7 @@
 import { z } from 'zod';
 import postgres from 'postgres';
 import { revalidatePath } from 'next/cache';
-// import { redirect } from 'next/navigation';
+import { redirect } from 'next/navigation';
 
 const sql = postgres(process.env.POSTGRES_URL!, { ssl: 'require' });
 
@@ -86,6 +86,7 @@ export async function createInvoice(prevState: State, formData: FormData): Promi
     }
 
     revalidatePath('/dashboard/invoices'); // Update route view (renew cache / prerender)
+    redirect('/dashboard/invoices');
 
     return {
         message: 'Invoice created successfully!',
@@ -94,7 +95,6 @@ export async function createInvoice(prevState: State, formData: FormData): Promi
         formData: null // Clear previous form data
     };
 
-    // redirect('/dashboard/invoices');
 };
 
 export async function updateInvoice(
@@ -152,9 +152,11 @@ export async function updateInvoice(
     // redirect('/dashboard/invoices');
 };
 
-export async function deleteInvoice(id: string): Promise<{ message: string, success: boolean }> {
+export async function deleteInvoice(id: string, prevState: State, formData: FormData): Promise<State> {
+    if (!id) return { message: 'Missing Invoice ID.', success: false };
+
     try {
-        await sql`DELETE FROM invoices WHERE id =${id}`;
+        await sql`DELETE FROM invoices WHERE id = ${id}`;
         revalidatePath('/dashboard/invoices');
         return { message: 'Invoice Deleted Successfully.', success: true };
